Rename 'components/MonitoringConsultationDeskReportChart.vue' to 'components/index/CardsReference/MonitoringConsultationDeskReportsNumber/Chart.vue'
